cleanup, moved query logic to util
